Refactored into conn, listener and common to be more navigable
